config: Add new `system_gitconfig` key to replace `omnibus_gitconfig`
In v15.2.0 we have split out the `gitaly['gitconfig']` key so that an
administrator can set the system-level Git configuration independently
of the Git configuration used by Gitaly. This was done to fix issues
with the deprecated `core.fsyncObjectFiles` configuration that needed to
exist on the system-level, but not on the Gitaly-level. Ultimately, the
configuration shared between both levels is almost always completely
distinct from each other.

As a consequence we need two separate keys to control either of both
levels. While this could theoretically be the the already-existing
`omnibus_gitconfig['system']` and `gitaly['gitconfig']`, this makes the
migration harder than necessary: the administrator would have to make
sure to always create the `gitaly['gitconfig']` key or otherwise we
might believe that there was nomigration yet.

Instead, add a new option `system_gitconfig` that only controls the Git
csnfiguration written at the system-level. Like this an administrator
would only have to rename the key to complete the migration in the
general case. And in a very limited set of cases they might also have to
set some specific keys for `gitaly['gitconfig']`.

Changelog: added