gitaly: Extract logic to convert Omnibus gitconfig
When splitting our the `gitaly['gitconfig']` configuration we have also
introduced a new format to specify Git configuration. To make the
migration less painful we have introduced a set of routines to convert
between the old and new format in the background.

We're about to fully deprecate the old `omnibus_gitconfig` key and will
thus introduce a new `system_gitconfig` key. In order to make things
easier for the administrator this new key will have the same format as
the `gitaly['gitconfig']` key. As a consequence there will be a second
callsite that needs to convert between the formats.

Prepare for this by pulling out the logic into a separate Gitconfig
helper so that it can easily be reused.