Convert Strings to Integer in Praefect's fallback config logic
Praefect's old configuration template automatically converted Strings
to Integers. While we've documented that the integer options should be
configured as Integer, the Strings have been accepted in the past as well.
This became apparent as GitLab Environment Toolkit is configuring the
database port as a string rather than an integer. While the issues is likely
not widespread due to the documentation configuring the relevant keys as
actual integers, let's handle this case to ensure the past config keeps
working if this has been done.