gitaly: Migrate away from `[gitlab-shell]` configuration section
Gitaly has deprecated the `[gitlab-shell]` section. This section
contained a single key `dir` from which two different paths were
derived:

    - The path of the file that contains the token required to connect
      to the GitLab API. This path is obtained by joining the directory
      path with ".gitlab_shell_secret". The setting has been replaced by
      a `[gitlab] secret_file` key.

    - The path of the directory that contains custom Git hooks. This
      path is obtained by joining the directory with "hooks". This
      setting has been replaced by a `[hooks] custom_hooks_dir` key.

Adjust the generated configuration file to make use of those new keys so
that Gitaly can eventually remove the old section. This refactoring is
not expected to cause any change in behaviour as we configure the new
settings to point to the exact same locations that the old setting did.