Fix Ruby builds when RUBY_VERSION is 3.2
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/7364
bumped the supported Ruby version from 3.2.2 to 3.2.3 and dropped the
Ruby interpreter patch that is no longer needed for 3.2.3. However, we
have scheduled pipelines that use `RUBY_VERSION` of 3.2, and the build
failed because 3.2 was presumed to be < 3.2.3. As a result, the Ruby
software package looked for the patch that was not present for Ruby
3.2.

Since we don't actually support Ruby 3.2.0 to 3.2.2 in the software
definition, we can relax the version check.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/438816

Changelog: fixed